Add to Download page a ZIP with all MSCZs.
When exporting the PDF, we now also export all MSCZs in a zip file.
This zip file is made available in the /Download page, next to the Jamicionário PDF.

In the future, we will instead prefer to host the .mscz files in a git submodule, and link the users to that repository.
There they will be able to get the data, but also to easily submit change proposals.