Files are served with only ASCII characters.
This avoids a known issue where files served from github pages get 404 if they have non-ASCII characters.
A solution could not be found, so we had to resort to limiting the characters in the filenames.

Only the filenames are changed.
No other data is changed: the titles are kept with their original characters, etc.

First the diacritics are removed, sanitizing most european letters into ASCII.
Then non-ASCII characters are removed, leaving only characters that are alphanumeric, space, or dashes.