Fix timeAgo pipe.
It was not calculating times well.
It was saying that the difference between the Jan-30th and Feb-2nd was "1 month ago" instead of "3 days ago".

It has been fixed, with imprecisions that are not important for now and were noted in comments.

Tested manually.
It needs automated tests. We added an entry in the TODO file.