Review version information.
Included version information in the Download page.

Added source URL to footer, because I think it's not very discoverable in the About page.

Changed the datetime property of the <time> element to ISO format.
The spec requires the value to be in one of a few specific formats, so we just use the most useful one.